Further styling of the inline popup and contents
Tag lineage uses negative margins to position itself at the outer edges
of the popup. Lineage colours and divider image added as well as simple
hover state.  A CSS3 arrow has been added to point to the selected
element, this might have to change to an image as its not possible to
add a shadow to this.  Buttons have been given a hover state.