Added a few range adjustments on mouseUp for double click word selection.
Double clicking a word at the end of a tag set the range.end AFTER the
tag in Firefox (and IE), but in Webkit it set the range inside the tag
which is the way it should work. This change moves the range inside the
tag in Firefox. IE does the same thing as Firefox however it also
selects the space before the next word and therefore this fix does not
work in IE as we don't know if user wanted to select the space or not…
This issue was noticed during testing of double clicking a word in a
strong tag.