Adding more functionality to Viper.createButton and Viper.createSubSection. Improved InlineToolbarPlugin's width calculation. Fixed CSS.
A button that toggles a sub section can now have a call back function
that will be called when the sub section is toggled. This allows plugin
to modify its parent elements with the new state.   Inline toolbar now
calculates its max width (including Sub sections) so that the popup does
not resize when a  sub section is toggled.   Fixed some of the CSS
issues when subsection was toggled (e.g. border radius). Note that the
grouped buttons still need styling.