Fixed InlineToolbar positioning and lineage in Chrome.
Webkit seems to get the range incorrectly when range is set on a node.
For example: <p>text</p><p>text</p> if the range.selectNode is called
for the first P then the next getCurrentRange call returns range start
as first P and range end as before the first character of the next 2nd P
tag. This causes range.getBoundingClientRect() to return the bottom
value as the 2nd P tag's bottom..