Styles cleanup and restructure
Sizing of all tools has been reduced to match that of the desired
desktop size. This means its going to be crazy small on the iPad until
the scaling code is updated.  CSS cleanup to remove some previous
commented out testing code. The media query has also been removed so a
single theme is currently in place across all devices.  CSS restructure
to support themes, this entails the separation of structural css from
colour rules. This will allow very easy creation and testing of a light
and dark theme (as the final desktop theme is yet to be decided). Themes
could also be used for device specific selection of theme (or user
selection as a future feature).  Minor tweaks to InlineToolbar sections
and subsections, as well as a start for button groupings.  InlineToolbar
arrow now works in IE. Although a lot of other button styles need work
for IE due to the above mentioned restructures.