Fixed issue with clicking Table's inline toolbar buttons that caused range to change.
Issue was that when the button is clicked (actually a mousedown event)
it updated the table structure moving the inline toolbar to another
location. And when the mouse is released (mouseup event) the
selectionChanged event fired on a different element causing toolbar to
be hidden or jump to the new element. Now the first selectionChanged
event after a button click is ignored.