Further highlight tweaks
Rather than a thick border which causes issues with different browsers
reporting different sizes for TD's I have reduced this to 1px and
increased the intensity of the box shadow. To aid darker tables I have
added a very transparent white inner shadow and background which is not
really visible in lighter tables.