Added a generic title element for use with subSection lists
The title element along with the subSection list and list item styles
are generic for re-use an any subSection that needs it. I see this
being re-used possibly in the custom styles lists.

The title is simply a div directly inside the subSection with a class
subSectionList-title. The UL or OL used afterwards needs the class
subSectionList. Another custom class should be added to the list for
any specific styles needed. For e.g. my Heading/Cell Lineage sample has
an extra class of cellLineage which help to position the text labels
and the abbreviation buttons.

Each LI in the lists needs the class subSectionList-item and any text
in a list item needs a span with class listItem-text.