Added accessibility option for VITP message box and required star for inputs
- Accessibility is now a state for message boxes, it has its own icon
of a small wheel-chair. The background colour is the standard
light-blue usually associated with accessibility.
- Testing an idea for inputs that are 'required'. Adding a simple class
of 'required' to the parent LABEL will automatically add a little star
into the text label and adjust the padding of the input accordingly.
This is done in such a way that if the required flag is added on the
fly the positioning of inputs and text labels will remain solid.