VCAP fix in Webkit for getElementAtCoords().
VCAP was disappearing when hovered over, when a table cell was at the end of
its content. getElementAtCoords() was returning undefined because it was
trying to access child node outside of its collection bounds.
Now if the start offset is beyond the length of the child nodes
collection, return the container itself.