MATRIX-702: Reinstate old method of enabling Keywords button as a backup
The way of determining whether to enable the Insert Keywords and
Snippets buttons was changed so that new Select2-based select boxes can
be used. But these currently require specific code in the WYSIWYG
content type and attribute, and therefore aren't available in Edit+.

In Edit+, Bodycopy Manager provides a backup for assets that contain a
bodycopy but it doesn't work all the time, and totally misses assets
which use a WYSIWYG attribute like Calendar Event and News Item.

I'm reinstating the old method of enabling the Keywords and Snippets
button - based on the "data-keywords" and "data-snippets" attributes of
the editable element - as a backup to the Select2-based method, so that
Edit+ can continue to use that.